Hermetic build (#846)
* Hermetic build


* Fix for review comments


---------

+++ scripts/lipo_vcpkg_macos.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
-# SPDX-FileCopyrightText: 2021-2026 Roy Shilkrot <roy.shil@gmail.com>
-# SPDX-FileCopyrightText: 2023-2026 Kaito Udagawa <umireon@kaito.tokyo>
+# SPDX-FileCopyrightText: 2025-2026 Kaito Udagawa <umireon@kaito.tokyo>
 #
-# SPDX-License-Identifier: GPL-3.0-or-later
+# SPDX-License-Identifier: Apache-2.0
 
 set -euo pipefail
+shopt -s nullglob
+
+if [[ "$#" -ne 3 ]]; then
+  echo "Usage: $0 <vcpkg_installed_arm64> <vcpkg_installed_x86_64> <vcpkg_installed_universal>"
+  exit 1
+fi
 
 rm -rf "$3"
 mkdir -p "$3"
@@ -24,7 +29,11 @@ cp -a "$VCPKG_INSTALLED_ARM64/include/" "$VCPKG_INSTALLED_UNIVERSAL/include/"
 cp -a "$VCPKG_INSTALLED_ARM64/lib/pkgconfig/" "$VCPKG_INSTALLED_UNIVERSAL/lib/pkgconfig/"
 cp -a "$VCPKG_INSTALLED_ARM64/share/" "$VCPKG_INSTALLED_UNIVERSAL/share/"
 
-for name in $(cd "$VCPKG_INSTALLED_ARM64/lib" && ls ./*.a); do
+pushd "$VCPKG_INSTALLED_ARM64/lib"
+libs=(*.a)
+popd
+
+for name in "${libs[@]}"; do
 	echo "Processing lib/$name"
 	lipo \
 		"$VCPKG_INSTALLED_ARM64/lib/$name" \
@@ -36,8 +45,13 @@ done
 if [[ -d "$VCPKG_INSTALLED_ARM64/debug" ]]; then
 	cp -a "$VCPKG_INSTALLED_ARM64/debug/lib/pkgconfig/" "$VCPKG_INSTALLED_UNIVERSAL/debug/lib/pkgconfig/"
 
-	for name in $(cd "$VCPKG_INSTALLED_ARM64/debug/lib" && ls ./*.a); do
+  pushd "$VCPKG_INSTALLED_ARM64/debug/lib"
+  debug_libs=(*.a)
+  popd
+
+	for name in "${debug_libs[@]}"; do
 		echo "Processing debug/lib/$name"
+
 		lipo \
 			"$VCPKG_INSTALLED_ARM64/debug/lib/$name" \
 			"$VCPKG_INSTALLED_X86_64/debug/lib/$name" \
